Make git signature status retrievable from Commit
We now include the status of the commit signature in the Commit
struct; which makes it easier to use throughout the codebase and
prevents unexpected results from one of the `Commits*()` methods.

It also moves the responsibility of actually verifying the commits
to where it should be; the daemon itself.